[HAL/NDK] - Make Vector parameter in HalEnableSystemInterrupt, HalDisableSystemInterrupt and HalBeginSystemInterrupt an ULONG, not an UCHAR [NDK] - 64bit fixes for HANDLE_TABLE, KPROCESS, SECTION_IMAGE_INFORMATION, MMADDRESS_LIST, MMVAD_FLAGS, MMVAD, MMVAD_LONG, MMVAD_SHORT, MEMORY_DESCRIPTOR, MEMORY_ALLOCATION_DESCRIPTOR, LdrVerifyMappedImageMatchesChecksum - KDPC_DATA::DpcQueueDepth is signed on amd64, unsigned on x86 [NTOSKRNL] - Fix hundreds of MSVC and amd64 warnings - add a pragma message to FstubFixupEfiPartition, since it looks broken - Move portable Ke constants from <arch>/cpu.c to krnlinit.c - Fixed a bug in amd64 KiGeneralProtectionFaultHandler
svn path=/trunk/; revision=53734

+++ reactos/include/ndk/ldrfuncs.h
@@ -105,7 +105,7 @@ BOOLEAN
 NTAPI
 LdrVerifyMappedImageMatchesChecksum(
     IN PVOID BaseAddress,
-    IN ULONG NumberOfBytes,
+    IN SIZE_T NumberOfBytes,
     IN ULONG FileLength
 );
